renamed component files for project's consistency
