+++ client/components/ViewChef.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
-import ViewChefItem from './view-chef-item';
+import ViewChefItem from './ViewChefItem';
 
 export default class ViewChef extends React.Component {
   constructor(props) {
